requirements.txt
streamlit>=1.30.0

pandas>=2.0.0
numpy>=1.24.0

sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

scikit-learn>=1.3.0
torch>=2.0.0

python-dotenv>=1.0.0
